Moved eventFire to separate component
